+++ test/dummy/app/models/locale_model.rb
@@ -1,0 +1,4 @@
+class LocaleModel < ApplicationRecord
+  validates :lo_value, locale: true
+  validates :nil_lo_value, locale: {allow_blank: true}
+end