+++ test/dummy/app/models/time_zone_model.rb
@@ -1,0 +1,4 @@
+class TimeZoneModel < ApplicationRecord
+  validates :tz_value, timezone: true
+  validates :nil_tz_value, timezone: { allow_blank: true }
+end