+++ test/dummy/db/migrate/20190721164936_create_time_zone_models.rb
@@ -1,0 +1,10 @@
+class CreateTimeZoneModels < ActiveRecord::Migration[5.2]
+  def change
+    create_table :time_zone_models do |t|
+      t.string :tz_value, null: false
+      t.string :nil_tz_value
+
+      t.timestamps
+    end
+  end
+end