+++ test/dummy/test/models/boolean_model_test.rb
@@ -1,7 +1,0 @@
-require 'test_helper'
-
-class BooleanModelTest < ActiveSupport::TestCase
-  # test "the truth" do
-  #   assert true
-  # end
-end